refactor bucket name generation

+++ pkg/provisioner/reconciler/claim-reconciler/reconiler.go
@@ -84,14 +84,14 @@ func (r *objectBucketClaimReconciler) Reconcile(request reconcile.Request) (reco
 		return handleErr("skipping provisioning for claim %s", obc.Name)
 	}
 
+	bucketName, err := util.GenerateBucketName(obc)
+	if err != nil {
+		return handleErr("error composing bucket name: %v", err)
+	}
+
 	class, err := util.StorageClassForClaim(obc, r.client, r.ctx)
 	if err != nil {
 		return handleErr("unable to get storage class: %v", err)
-	}
-
-	bucketName := obc.Spec.BucketName
-	if bucketName == "" {
-		bucketName = util.GenerateBucketName(obc.Spec.GeneratBucketName)
 	}
 
 	options := &api.BucketOptions{
@@ -134,10 +134,16 @@ func (r *objectBucketClaimReconciler) handelReconcile(options *api.BucketOptions
 	// so we can start fresh in the next iteration
 	defer func() {
 		if err != nil {
-			_ = r.provisioner.Delete(ob)
-			_ = r.client.Delete(context.Background(), ob)
-			_ = r.client.Delete(context.Background(), secret)
-			_ = r.client.Delete(context.Background(), configMap)
+			if ob != nil {
+				_ = r.provisioner.Delete(ob)
+				_ = r.client.Delete(context.Background(), ob)
+			}
+			if secret != nil {
+				_ = r.client.Delete(context.Background(), secret)
+			}
+			if configMap != nil {
+				_ = r.client.Delete(context.Background(), configMap)
+			}
 		}
 	}()
